+++ custom_components/fritzbox_tools/binary_sensor.py
@@ -17,7 +17,7 @@ except ImportError:
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.typing import HomeAssistantType
 
-from . import DATA_FRITZ_TOOLS_INSTANCE, DOMAIN
+from .const import DATA_FRITZ_TOOLS_INSTANCE, DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
